friend management

+++ code/elife/user-server/src/main/java/user/entity/Manager.java
@@ -31,10 +31,12 @@ public class Manager implements Serializable {
     private Integer role;
 
     @Column(name = "community_id")
-    private Integer communityId;
+    private Long communityId;
 
-    public Manager(){}
-    public Manager(String username, String password, String phone, String email, Integer role, Integer communityId){
+    public Manager() {
+    }
+
+    public Manager(String username, String password, String phone, String email, Integer role, Long communityId) {
         this.username = username;
         this.password = password;
         this.phone = phone;
@@ -83,11 +85,11 @@ public class Manager implements Serializable {
         this.role = role;
     }
 
-    public Integer getCommunityId() {
+    public Long getCommunityId() {
         return communityId;
     }
 
-    public void setCommunityId(Integer communityId) {
+    public void setCommunityId(Long communityId) {
         this.communityId = communityId;
     }
